progress: fix progress does not work with CJK
The display width of a rune may have 0, 1 or 2. Especially for CJK characters,
nearly all CJK characters are 2 wide.

The code before this patch assumes the width of all runes is 1, it causes the
width of a text which have CJK characters has been wrongly calculated. Thus,
the progress status text may overflow into next line on the terminal.

This patch fix this issue by using runewidth to detect the length of each rune.

Fixes: https://bugs.launchpad.net/snapd/+bug/1952500

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>